Bug fix shmemx_memcpy, faster code, smaller size
This routine handles aligned memory copying in a high performance manner. If
the source and desination arrays have byte offsets but are dword aligned at
some point, the routine copies the initial offset amounts, performs dword
copies in the middle, and then copies the remaining bytes. If the arrays are
misaligned, it performs a slow copy.