Portability improvements for test codes
The test codes call ctimer routines, which has a generic API rather than
specific to Epiphany.  The remnants of the phost_printf calls were removed.
The global variables and arrays have been moved to static and properly
initialized.  The previous array initialization would generally only work for
this particular SHMEM implementation because the initial value was 0, but could
be different on other implementations. The internal loops for timing have been
increased to get a better sample size.  Consequently, some codes run longer and
the ctimer_start is moved in the internal loop so that it's less likely to
underflow (32-bit ctimer specific to Epiphany). The codes should now be mostly
portable if one provides ctimer routines.